// src/shared/interfaces/modules/user/IUser.ts
export interface IUser {
  id: string;
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  profileImage: string;
  phone: string;
  isVerified: boolean;
  address: string;
  district: string;
  city: string;
  state: string;
  zip_code: string;
  planUser: string;
  planStatus: string;
  customerId: string;
}